tidy up: added new files
