+++ pyrateoptics/optimize/optimize.py
@@ -24,262 +24,7 @@ along with this program; if not, write to the Free Software
 Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 """
 
-import numpy as np
-import math
-
 from ..core.log import BaseLogger
-
-class OptimizableVariable(BaseLogger):
-    """
-    Class that contains an optimizable variable. Used to get a pointer on a variable.
-    The value is not constrained to float. Also other dependent variables are possible to define.
-    """
-    def __init__(self, variable_type="fixed", **kwargs):
-
-        super(OptimizableVariable, self).__init__(name=kwargs.pop('name', ''), **kwargs)
-
-        """
-        kwargs depend on type
-        "variable" is value=value
-        "pickup" gets function=f, args=tuple of optimizablevariables
-        "external" gets function=f, args=tuple of external standard variables 
-        (or values)
-
-        :param variable_type (string): which kind of variable do we have? 
-        valid choices are: 'fixed', 'variable', 'pickup', 'external'
-
-        :param kwargs (keyword arguments): used to fill up parameters 
-        dictionary for variable.
-
-        kwargs:
-
-        type:        kwargs:
-        'fixed'         value=1.0 or value='hello'
-        'variable'    value=1.0, or value='hello'
-        'pickup'        function=f, args=(other_optvar1, other_optvar2, ...)
-        'external'        function=f, args=(other_externalvar/value, ...)
-
-        Notice: f must take exactly as many arguments as args=... long is. 
-        The only constraint on f is that it has to convert the variables into 
-        some final float variable. If this is not the case some
-        low-level optimizer might break down.
-
-
-        """
-
-        self.evaldict = {
-                    "fixed"    : self.eval_fixed,
-                    "variable" : self.eval_variable,
-                    "pickup"   : self.eval_pickup,
-                    "external" : self.eval_external
-                    }
-
-
-        self.changetype(variable_type)
-        self.parameters = kwargs
-
-        self.initial_value = self.evaluate()
-        self.set_interval(None, None)
-
-
-    def __str__(self, *args, **kwargs):
-        return self.name + "('" + self.var_type + "') = " + str(self.parameters)
-
-    def getVarType(self):
-        return self.__var_type.lower()
-    
-    def setVarType(self, vtype):
-        self.__var_type = vtype.lower()
-
-    var_type = property(fget=getVarType, fset=setVarType)
-
-    def set_interval(self, left=None, right=None):
-        """
-        Interval transform from finite to infinite and back.
-        The reason for this is that the optimizer backend has
-        only to deal with infinite range variables and the user
-        can transparently change the interval. The idea and
-        verification is (up to a slight change in the 
-        transformation) taken from:
-        M. Roeber, "Multikriterielle Optimierungsverfahren fuer
-        rechenzeitintensive technische Aufgabenstellungen",
-        Diploma Thesis, Technische Universitaet Chemnitz, 
-        2010-03-31
-        """
-        # TODO: fine-tune for one-sided intervals
-        if left is None and right is None:
-            self.transform = lambda x: x
-            self.inv_transform = lambda x: x
-        else:
-            self.inv_transform = lambda x: left + (right - left)/(1. + math.exp(-x/math.fabs(right - left)))
-            self.transform = lambda x: math.log((-x + left)/(x - right))*math.fabs(left - right)
-
-    def changetype(self, vtype, **kwargs):
-        self.var_type = vtype
-        self.evalfunc = self.evaldict[self.var_type]
-        try:        
-            parameter_backup = self.parameters
-        except:
-            parameter_backup = {"value":None}
-        self.parameters = kwargs
-        if vtype.lower() in ["variable", "fixed"] \
-            and self.var_type in ["variable", "fixed"] \
-            and parameter_backup["value"] != None:
-            self.parameters["value"] = parameter_backup["value"]
-
-    def setvalue(self, value):
-        # TODO: overload assign operator
-        if self.var_type == "variable" or self.var_type == "fixed":
-            self.parameters["value"] = value
-
-    def eval_fixed(self):
-        # if type = variable then give only access to value
-        return self.parameters["value"]
-
-    def eval_variable(self):
-        # if type = variable then give only access to value
-        return self.parameters["value"]
-
-    def eval_pickup(self):
-        # if type = pickup then pack all arguments into one tuple
-        # and put it into the userdefined function
-        # evaluate the result
-        arguments_for_function_eval = (argfunc.evaluate() for argfunc in self.parameters["args"])
-        return self.parameters["function"](*arguments_for_function_eval)
-
-    def eval_external(self):
-        # same as for solve except that there are no further OptimizableVariables to be considered
-        return self.parameters["function"](*self.parameters["args"])
-
-    def evaluate(self):
-        # notice: evaluation code is not limited to floats
-        # do not use if-else construct because for many cases dict lookup is faster
-
-        return self.evalfunc()
-        
-    def __call__(self):
-        return self.evaluate()
-        
-    def evaluate_transformed(self):
-        """
-        Transform variable value from finite interval 
-        to infinite IR before evaluation
-        """
-        return self.transform(self.evaluate())
-        
-    def setvalue_transformed(self, value_transformed):
-        """
-        Transform value back from infinite IR before setting value
-        """
-        value = self.inv_transform(value_transformed)
-        self.setvalue(value)
-        
-class ClassWithOptimizableVariables(BaseLogger):
-    """
-    Implementation of some class with optimizable variables with the help 
-    of a dictionary. This class is also able to collect the variables and 
-    their values from its subclasses per recursion.
-    """
-    def __init__(self, name = "", **kwargs):
-        """
-        Initialize with empty dict.
-        """
-        super(ClassWithOptimizableVariables, self).__init__(name=name, **kwargs)
-
-        self.list_observers = [] 
-        # for the optimizable variable class it is useful to have some observer links
-        # they get informed if variables change their values
-
-    def appendObservers(self, obslist):
-        self.list_observers += obslist
-
-    def informObservers(self):
-        for obs in self.list_observers:
-            obs.informAboutUpdate()
-
-                
-    def getAllVariables(self):
-
-        def addOptimizableVariablesToList(var, dictOfOptVars = {}, idlist=[], keystring="", reducedkeystring=""):
-            """
-            Accumulates optimizable variables in var and its linked objects.
-            Ignores ring-links and double links.       
-            
-            @param var: object to evaluate (object)
-            @param dictOfOptVars: optimizable variables found so far (dict of objects)
-            @param idlist: ids of objects already evaluated (list of int)
-            """ 
-            
-            if id(var) not in idlist:
-                idlist.append(id(var))
-    
-                if isinstance(var, ClassWithOptimizableVariables):
-                    for (k, v) in var.__dict__.items():
-                        newkeystring = keystring + "(" + var.name + ")." + str(k)
-                        newredkeystring = reducedkeystring + var.name + "."
-                        dictOfOptVars, idlist = addOptimizableVariablesToList(v, dictOfOptVars, idlist, newkeystring, newredkeystring) 
-                elif isinstance(var, dict):
-                    for (k, v) in var.items():
-                        newkeystring = keystring + "[" + str(k) + "]" 
-                        dictOfOptVars, idlist = addOptimizableVariablesToList(v, dictOfOptVars, idlist, newkeystring, reducedkeystring)                        
-    
-                elif isinstance(var, list) or isinstance(var, tuple):
-                    for (ind, v) in enumerate(var):
-                        newkeystring = keystring + "[" + str(ind) + "]" 
-                        dictOfOptVars, idlist = addOptimizableVariablesToList(v, dictOfOptVars, idlist, newkeystring, reducedkeystring)
-    
-                elif isinstance(var, OptimizableVariable):
-                    newkeystring = keystring + "(" + var.name + ")"
-                    newreducedkeystring = reducedkeystring + var.name
-                    #dictOfOptVars.append((var, newkeystring, newreducedkeystring))
-                    dictOfOptVars[newreducedkeystring] = (var, newkeystring)
-    
-            return dictOfOptVars, idlist
-
-
-
-        (dict_opt_vars, idlist) = addOptimizableVariablesToList(self)
-        return dict_opt_vars
-
-
-    def getAllValues(self):
-        """
-        For fast evaluation of value vector
-        """
-        return np.array([a.evaluate() for (a, b) in self.getAllVariables().values()])
-
-    def getActiveVariables(self):
-        """
-        Returns a list of active variables the names are lost
-        but it does not matter since the variable references are still in the
-        original dictionary in the class
-        """
-        return [x for (x, y) in self.getAllVariables().values() if x.var_type == "variable"]
-
-
-    def getActiveValues(self):
-        """
-        Function to get all values into one large np.array.
-        """
-        return np.array([a() for a in self.getActiveVariables()])
-
-    def setActiveValues(self, x):
-        """
-        Function to set all values of active variables to the values in the large np.array x.
-        """
-        for i, var in enumerate(self.getActiveVariables()):
-            var.setvalue(x[i])
-
-    def getActiveTransformedValues(self):
-        return np.array([a.evaluate_transformed() for a in self.getActiveVariables()])
-
-    def setActiveTransformedValues(self, x):
-        """
-        Function to set all values of active variables to the values in the large np.array x.
-        """
-        for i, var in enumerate(self.getActiveVariables()):
-            var.setvalue_transformed(x[i])
-
 
 class Optimizer(BaseLogger):
     '''
